Refactor record
